// src/generated/models/Pm_holidaiesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Pm_holidaiesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Pm_holidaiesstatecode = keyof typeof Pm_holidaiesstatecode;
export const Pm_holidaiesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Pm_holidaiesstatuscode = keyof typeof Pm_holidaiesstatuscode;

export interface Pm_holidaiesBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  pm_country?: string;
  pm_holidaydate?: string;
  pm_holidayid: string;
  pm_holidayname?: string;
  pm_isfixeddate?: boolean;
  pm_notes?: string;
  pm_year?: number;
  statecode: Pm_holidaiesstatecode;
  statuscode?: Pm_holidaiesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Pm_holidaies extends Pm_holidaiesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  pm_isfixeddatename?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}